MOD: refactor user account list with widget
